app/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
gunicorn
gTTS==2.5.1
googletrans==4.0.0rc1
pydub